mcp-server/requirements.txt
mcp>=0.9.0
anthropic>=0.34.0
yahooquery>=2.3.7
exa-py>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

